rag-test-python/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
tiktoken>=0.9.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.32.0
colorama>=0.4.6
